Fix lib name conflict: BLEPeripheral => BLEPeripheralObserver.
