+++ examples/HID/HID_volume/HID_volume.ino
@@ -1,7 +1,7 @@
 // Copyright (c) Sandeep Mistry. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
-// Import libraries (BLEPeripheral depends on SPI)
+// Import libraries (BLEPeripheralObserver depends on SPI)
 #include <SPI.h>
 #include <BLEHIDPeripheral.h>
 #include <BLEMultimedia.h>
