+++ examples/led/led.ino
@@ -1,15 +1,15 @@
 // Copyright (c) Sandeep Mistry. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
-// Import libraries (BLEPeripheral depends on SPI)
+// Import libraries (BLEPeripheralObserver depends on SPI)
 #include <SPI.h>
-#include <BLEPeripheral.h>
+#include <BLEPeripheralObserver.h>
 
 // LED pin
 #define LED_PIN   3
 
-//custom boards may override default pin definitions with BLEPeripheral(PIN_REQ, PIN_RDY, PIN_RST)
-BLEPeripheral                    blePeripheral                            = BLEPeripheral();
+//custom boards may override default pin definitions with BLEPeripheralObserver(PIN_REQ, PIN_RDY, PIN_RST)
+BLEPeripheralObserver                    blePeriphObserv                            = BLEPeripheralObserver();
 
 // create service
 BLEService               ledService           = BLEService("19b10000e8f2537e4f6cd104768a1214");
@@ -27,21 +27,21 @@ void setup() {
   pinMode(LED_PIN, OUTPUT);
 
   // set advertised local name and service UUID
-  blePeripheral.setLocalName("LED");
-  blePeripheral.setAdvertisedServiceUuid(ledService.uuid());
+  blePeriphObserv.setLocalName("LED");
+  blePeriphObserv.setAdvertisedServiceUuid(ledService.uuid());
 
   // add service and characteristic
-  blePeripheral.addAttribute(ledService);
-  blePeripheral.addAttribute(switchCharacteristic);
+  blePeriphObserv.addAttribute(ledService);
+  blePeriphObserv.addAttribute(switchCharacteristic);
 
   // begin initialization
-  blePeripheral.begin();
+  blePeriphObserv.begin();
 
   Serial.println(F("BLE LED Peripheral"));
 }
 
 void loop() {
-  BLECentral central = blePeripheral.central();
+  BLECentral central = blePeriphObserv.central();
 
   if (central) {
     // central connected to peripheral
