+++ examples/led_switch/led_switch.ino
@@ -1,16 +1,16 @@
 // Copyright (c) Sandeep Mistry. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
-// Import libraries (BLEPeripheral depends on SPI)
+// Import libraries (BLEPeripheralObserver depends on SPI)
 #include <SPI.h>
-#include <BLEPeripheral.h>
+#include <BLEPeripheralObserver.h>
 
 // LED and button pin
 #define LED_PIN     3
 #define BUTTON_PIN  4
 
-//custom boards may override default pin definitions with BLEPeripheral(PIN_REQ, PIN_RDY, PIN_RST)
-BLEPeripheral                    blePeripheral                            = BLEPeripheral();
+//custom boards may override default pin definitions with BLEPeripheralObserver(PIN_REQ, PIN_RDY, PIN_RST)
+BLEPeripheralObserver                    blePeriphObserv                            = BLEPeripheralObserver();
 
 // create service
 BLEService               ledService           = BLEService("19b10010e8f2537e4f6cd104768a1214");
@@ -30,23 +30,23 @@ void setup() {
   pinMode(BUTTON_PIN, INPUT);
 
   // set advertised local name and service UUID
-  blePeripheral.setLocalName("LED Switch");
-  blePeripheral.setAdvertisedServiceUuid(ledService.uuid());
+  blePeriphObserv.setLocalName("LED Switch");
+  blePeriphObserv.setAdvertisedServiceUuid(ledService.uuid());
 
   // add service and characteristics
-  blePeripheral.addAttribute(ledService);
-  blePeripheral.addAttribute(switchCharacteristic);
-  blePeripheral.addAttribute(buttonCharacteristic);
+  blePeriphObserv.addAttribute(ledService);
+  blePeriphObserv.addAttribute(switchCharacteristic);
+  blePeriphObserv.addAttribute(buttonCharacteristic);
 
   // begin initialization
-  blePeripheral.begin();
+  blePeriphObserv.begin();
 
   Serial.println(F("BLE LED Switch Peripheral"));
 }
 
 void loop() {
   // poll peripheral
-  blePeripheral.poll();
+  blePeriphObserv.poll();
 
   // read the current button pin state
   char buttonValue = digitalRead(BUTTON_PIN);
