+++ examples/test/test.ino
@@ -7,12 +7,12 @@
 #include <MemoryFree.h>
 #endif
 
-// Import libraries (BLEPeripheral depends on SPI)
+// Import libraries (BLEPeripheralObserver depends on SPI)
 #include <SPI.h>
-#include <BLEPeripheral.h>
+#include <BLEPeripheralObserver.h>
 
-//custom boards may override default pin definitions with BLEPeripheral(PIN_REQ, PIN_RDY, PIN_RST)
-BLEPeripheral                    blePeripheral                            = BLEPeripheral();
+//custom boards may override default pin definitions with BLEPeripheralObserver(PIN_REQ, PIN_RDY, PIN_RST)
+BLEPeripheralObserver                    blePeriphObserv                            = BLEPeripheralObserver();
 
 // create service
 BLEService                       testService         = BLEService("fff0");
@@ -30,26 +30,26 @@ void setup() {
   delay(5000);  //5 seconds delay for enabling to see the start up comments on the serial board
 #endif
 
-  blePeripheral.setLocalName("test");
+  blePeriphObserv.setLocalName("test");
 #if 1
-  blePeripheral.setAdvertisedServiceUuid(testService.uuid());
+  blePeriphObserv.setAdvertisedServiceUuid(testService.uuid());
 #else
   const char manufacturerData[4] = {0x12, 0x34, 0x56, 0x78};
-  blePeripheral.setManufacturerData(manufacturerData, sizeof(manufacturerData));
+  blePeriphObserv.setManufacturerData(manufacturerData, sizeof(manufacturerData));
 #endif
 
   // set device name and appearance
-  blePeripheral.setDeviceName("Test");
-  blePeripheral.setAppearance(0x0080);
+  blePeriphObserv.setDeviceName("Test");
+  blePeriphObserv.setAppearance(0x0080);
 
   // add service, characteristic, and decriptor to peripheral
-  blePeripheral.addAttribute(testService);
-  blePeripheral.addAttribute(testCharacteristic);
-  blePeripheral.addAttribute(testDescriptor);
+  blePeriphObserv.addAttribute(testService);
+  blePeriphObserv.addAttribute(testCharacteristic);
+  blePeriphObserv.addAttribute(testDescriptor);
 
   // assign event handlers for connected, disconnected to peripheral
-  blePeripheral.setEventHandler(BLEConnected, blePeripheralConnectHandler);
-  blePeripheral.setEventHandler(BLEDisconnected, blePeripheralDisconnectHandler);
+  blePeriphObserv.setEventHandler(BLEConnected, blePeripheralConnectHandler);
+  blePeriphObserv.setEventHandler(BLEDisconnected, blePeripheralDisconnectHandler);
 
   // assign event handlers for characteristic
   testCharacteristic.setEventHandler(BLEWritten, characteristicWritten);
@@ -60,7 +60,7 @@ void setup() {
   testCharacteristic.setValue(0);
 
   // begin initialization
-  blePeripheral.begin();
+  blePeriphObserv.begin();
 
   Serial.println(F("BLE Peripheral"));
 
@@ -71,7 +71,7 @@ void setup() {
 }
 
 void loop() {
-  BLECentral central = blePeripheral.central();
+  BLECentral central = blePeriphObserv.central();
 
   if (central) {
     // central connected to peripheral
