+++ src/BLELocalAttribute.h
@@ -8,7 +8,7 @@
 
 class BLELocalAttribute : public BLEAttribute
 {
-  friend class BLEPeripheral;
+  friend class BLEPeripheralObserver;
 
   public:
     BLELocalAttribute(const char* uuid, enum BLEAttributeType type);
